Be more platform-agnostic for `hlint`
Turns out `make` doesn't exist on the Windows environment of TravisCI.
It has `cmake` instead. Also, the macOS environment of TravisCI doesn't
have the same version of `mktemp` or `rm` that exists on the Linux
environment. It is probably using the BSD versions of those while we
expected the GNU versions.

We write things more agnostic by taking in the build and bin
directories. This way, we can call the script with arguments instead of
expecting it to know more information than it should. It was dubious of
us to assume the path to `bin` would always work anyway.

To address the Windows issue, we call the script directly to install
`hlint` and run `hlint` manually once it's installed. This isn't ideal
because it means that we have two workflows: one using `make` the other
orchestrating things by hand. Hopefully, we can address these issues
better in the future.