Fix environment variables
Looks like each line in this YAML string is run as its own shell
process. So maybe the variables aren't available directly to the script?
It's unclear why this didn't work but a few lines above we had `URL`
available in the environment.

In any case, we try putting it all on one line so it should hopefully be
part of the environment the script has available.