Create issue templates
We want a better experience for people using the issue tracker.

This is the first step in making things better for people reporting
issues. We want to direct people toward creating an issue that provides
the right information.

For now, we offer two options: bug reports, and concrete compiler
proposals. If we find that we want a different set of options, or a
different template for others to use, we can change this at a later
date.