Add `contact_links` for general questions
We want to point people to Discourse for general questions.

We've been pushing for people to use Discourse for ideas or questions
for quite a while now. In an effort to make things a bit easier on
everyone, we want to have a `contact_link` that nudges people there from
the get-go.

Hopefully, this is an improvement on the current state of affairs.