Create issue templates (#3853)
* Create issue templates

We want a better experience for people using the issue tracker.

This is the first step in making things better for people reporting
issues. We want to direct people toward creating an issue that provides
the right information.

For now, we offer two options: bug reports, and concrete compiler
proposals. If we find that we want a different set of options, or a
different template for others to use, we can change this at a later
date.

* Remove erroneous file

This file was accidentally added.

* Add `contact_links` for general questions

We want to point people to Discourse for general questions.

We've been pushing for people to use Discourse for ideas or questions
for quite a while now. In an effort to make things a bit easier on
everyone, we want to have a `contact_link` that nudges people there from
the get-go.

Hopefully, this is an improvement on the current state of affairs.

* Update .github/ISSUE_TEMPLATE/bug_report.md