Check role declarations against inferred roles
Fixes #3867. The main change here is that we are now storing checked
roles in the `types` field of the Environment for `data` and
`foreign import data` declarations, via the additions to the `TypeKind`
type. Because we are storing roles in the environment, we no longer need
to infer roles for any given type more than once. As such, `checkRoles`
now works in a monad with `MonadState CheckState`, and when we come
across another type while inferring roles, we can simply look its roles
up from the Environment. The BindingGroups desugaring step ensures
that types are in topological order, so this is safe: any mentioned
types' roles will already have been checked and stored in the
Environment.

One case that we do not yet handle is types which cyclically depend on
each other, e.g.

   data F a = F (G a)
   data G a = G (F a)

Rather than crashing the compiler, these will now be incorrectly marked
Phantom. I would like to address this in a future PR.