Bump version for 0.14.0-rc2
The rc1 I already tagged is no good, because I forgot to update the
version. That is particularly important for this release, because tools
like psa need to be able to know what version of the CLI they have
encountered in order to know whether to read from stdout or stderr if
they want to maintain compatibility with both 0.13.x and 0.14.x.