Fix #3953
This commit resolves a grammar conflict between the rule allowing
negative numbers to be atomic binders (so that, e.g., `foo -1 = True` is
a valid partial function declaration) and the rule allowing the minus
sign to be an operator separating binders.

The total number of conflicts in the grammar increases by one with this
commit, but what is actually happening is that a state that previously
had one conflict is now conflict-free, and two states that already had
conflicts involving the token '-' now have one extra conflict each. The
extra conflicts aren't conceptually new, though; those states correspond
to productions where a rule containing '-' was duplicated with OP_NEGATE
replacing '-'. So despite the %expect number increasing, I consider this
a net decrease in the conflicted-ness of the grammar.