Use type annotation hint only when needed
This commit changes the criteria for when the message

  The instance head contains unknown type variables. Consider adding a
  type annotation.

appears in a NoInstanceFound error. Instead of using it whenever an
unknown type variable appears in an instance head, now the compiler
adds this hint only when every covering set of type arguments of the
instance head contains an unknown type variable. (Covering sets are
determined by functional dependencies, if present; otherwise, the entire
set of type arguments is the one and only covering set.)

The rationale is that if any covering set is already completely known,
then the availability of an appropriate instance will not be changed by
adding more type annotations.