Allow fixity, kind, role declarations in REPL
It seems that advances in parsing and/or interpretation have made the
artificial restriction that used to exist on declarations in the REPL no
longer necessary. Fixity, kind, and role declarations are all tested and
working, and the only other types of declarations that exist in the AST
at this time are imports (which are specially handled in an earlier
parser), BoundValueDeclarations (which can't be created at the top
level), and binding groups (which are only created in desugaring).