Fix wildly off kind unification positions
Fixes #4019

This scrubs the kind-checker's error position context when invoking
kind unification from type-checking routines. When the kind-checker
errors, we want to preserve the context of the unification constraint.
Unfortunately, we don't have that position on hand since it's maintained
implicitly as part of the type-checker stack, so this is a somewhat
indirect solution.