Drop libtinfo dependency
Resolves #3696. (Note that #3696 says "use static linking in binary
distributions", and this doesn't actually do that as there are still a
number of other dynamic library dependencies. However all of the
remaining ones are very common and likely to already exist on the user's
system, so that users hopefully shouldn't need to worry about dynamic
library dependencies at all after this change, which hopefully does
resolve the issue in practice.)

We regularly hear from beginners who are having difficulty installing
the compiler because of the libtinfo dependency, so let's drop it.

The REPL experience is only very marginally degraded by this change, and
I suspect a large chunk of users don't even use the REPL anyway -
certainly CI environments, where the libtinfo dependency is often an
issue, don't care about the REPL. If people do want the terminfo-based
REPL they can still build their own version of the compiler with the
haskeline `terminfo` flag enabled.

See https://github.com/purescript/purescript/issues/3696#issuecomment-657282303
for a comparison.