Do less work in test initialization
The goal of this commit is to minimize the amount of time it takes for
Hspec to get from the starting line to a complete spec tree, and thus to
improve the time it takes to run individual or small numbers of tests
from the suite. First, Hspec `beforeAll`/`beforeAllWith` hooks are used
to defer some work until Hspec actually runs the tests that require that
work. Second, the `npm install` and `bower install` pre-test steps are
throttled to run once a day, provided that the output directories
already exist and that their input files haven't changed.