Improve apartness checking
Instances in an instance chain that don't match the constraint being
searched must be proven apart from the constraint for the search to
continue to the next instance in the chain. This commit tweaks the
apartness checker to fix a variety of incorrect apartness judgments,
and also adds a hint to the NoInstanceFound error message if ambiguous
instances (neither matching nor apart) excluded additional instances in
a chain from consideration.