Rewrite addRoles so that it's stack-safe
This commit also throws an internalError if two situations where it 
didn't previously. First, if the number of roles don't match the number 
of type vars. Second, if we somehow attempt to add another role to a 
type var that already has one. None of these should be possible by this 
point, but should it ever happen, it'll help us track down what went 
wrong.