Display roles in HTML docs (#4121)
Including roles in markdown docs was not added in this PR. Markdown docs currently
render their docs a specific way that would require reworking that part to fix. Most
people don't use Markdown docs now that HTML docs are the norm, so this shouldn't
be an issue.

FFI declarations are converted to DataDeclarations so that we can still show their roles
in their corresponding docs. See below for a longer explanation.

Finally, all comments on kind signatures, declarations, and role declarations are merged
together in the final docs.

* Include role declarations in list of ordered exports

* Insert roles in doc decl augmentation pass

* Render role annotations in HTML docs but not markdown docs

* Add test: verify expected role declarations match actual ones

* Swap parameter order for renderRole

* Link to roles page on doc repo when role annotation is clicked

* Ensure inferred types are Text, not String

* Don't mention values as some types are intentionally uninhabited

* Ensure role page is opened in new window/tab

* Render role as italics, sans-serif font with no underline

* Make role links unbolded

* Use clearer error message if add a 2nd role declaration to data decl

* Add/refactor doc-comment merging tests to include role declarations

* Ignore rather than crash when attempting to add role declarations to FFI

* Rewrite addRoles so that it's stack-safe

This commit also throws an internalError if two situations where it 
didn't previously. First, if the number of roles don't match the number 
of type vars. Second, if we somehow attempt to add another role to a 
type var that already has one. None of these should be possible by this 
point, but should it ever happen, it'll help us track down what went 
wrong.

* Throw an error if we get an unexpected value for role annotation word

* Include `RoleDeclaration`s in `declName`

* Use one renderType with two faces to prevent code from unsyncing

* Convert FFI declarations into data declaration

FFI declarations (i.e. ExternDataDeclaration) are rendered using only 
their kind signature. Since roles are rendered by annotating the type 
parameters of a normal signature, we need to convert the FFI 
declarations to a data declaration, so that we can annotate that data 
declaration's type parameters with their corresponding roles. However, 
the FFI declaration does not have type parameters. Thus, we infer how 
many type parameters it implicitly has based on its kind signature and 
then generate a name for those type parameters. The generated names are 
easily distinguishable from developer-defined ones (e.g. `data Foo a b`) 
because they are generated via `t` and an 0-based index. When most 
developers typically use `a`, `b`, `f`, or `m` as the names of their 
type parameters (or more descriptive names like `container`), this 
different pattern will stand out and indicate that the type is an FFI 
data type

* Insert inferred roles if none were explicitly declared

* Insert inferred roles for data declarations

This also inserts inferred roles for FFI declarations due to coming 
after the FFI conversion pass

* Add test checking for role annotations for FFI data types

* Use rhendric's rendering for role explanation (argument version)

* Rename E to FFI to clarify type of declaration

* Add tests for higher-kinded type parameters

* Add tests for FFI declarations' kind signatures

* Verify that FFI declarations with redundant parens are accounted for

* Test all possible variations of doc comment merging

* Update changelog to clarify the FFI declaration rendering situation

* Add note: intentionally not dropping sort ann for explicit kind sigs

* Fix typo

Co-authored-by: Cyril <sobierajewicz.cyril@gmail.com>

* Re-enable FFI role declaration tests

Co-authored-by: Cyril <sobierajewicz.cyril@gmail.com>