Build the entire latest package set in CI (#4217)
* Build the entire latest package set in CI

The package set is only built in Linux, as that tends to be the OS that
otherwise finishes first, and there's no need to run this test three
times. The package set is *not* built if there are any files in
CHANGELOG.d starting with `breaking_`, indicating that there are
unreleased breaking changes.