Rewrite `Partial` optimization to be cleaner (#4208)
This feature shrinks the generated JS code for declarations that use
empty type classes, such as `Partial`, but is otherwise not expected to
have user-visible consequences.

A previous issue involving `Partial` resulted in an ad-hoc optimization
pass added to the CoreFn phase. This pass consumed code generated by the
exhaustiveness checker, which used `UnusedIdent` in a questionable
manner, and ran a rewrite on it that assumed that the only part of the
compile that would use `UnusedIdent` in such a questionable manner was
necessarily the exhaustiveness checker.

This commit:

* simplifies the code generated by the exhaustiveness checker for
  partial matches, in particular removing any use of `UnusedIdent`s;

* removes the offending CoreFn optimization pass;

* solves the problem originally motivating said optimization pass by
  using an `UnusedIdent` for the generated parameter when type-checking
  a value constrained by an empty type class (such parameters,
  importantly for our sanity, are actually unused);

* and backs out a related hack in TCO that worked around such generated
  parameters being present but called without an argument.

The intended result is simpler compiler code; happy side effects include
simpler error messages involving partial pattern matches (seen here in a
few golden test output changes) and simpler generated code (fewer unused
function parameters, fewer local variables in some TCO-triggering
functions).