Disable constraints in FFI (#4240)
* Error when constraints are found in FFI values

For example...
```
foreign import foo :: forall a. Show a => a -> String
```

Constraints are not allowed because one can implicitly rely upon
the dictionary's representation at runtime.

* Add test for FFI constraint workaround

In other words, type class dictionary members (e.g. `show` below)
must be passed explicitly into the FFI function for one to use
them in FFI.
```
foreign import fooImpl :: forall a. (a -> String) -> a -> String

foo :: forall a. Show a => a -> String
foo a = fooImpl show a
```

Note: this test may break if we change the Show instances
for any of the used types in `purescript-prelude`.

* Rename test files to drop 'Deprecated'

* Add changelog entry