Lazy initialization for recursive bindings
This commit adds src/Language/PureScript/CoreFn/Laziness.hs, a
transformation on CoreFn that introduces call-by-need semantics to
bindings in recursive binding groups. This is the only way to restore
soundness to those groups without forbidding many programs we've
previously allowed, as historically any reference to another identifier
within the same binding group has been permitted as long as it occurs
within a lambda, regardless of whether that lambda is reduced in the
course of initialization.

The rest of PureScript remains call-by-value. Also, the imperfect
prohibitions on cyclic references enforced during desugaring and type
checking remain in place; consequently, this change does not alter the
set of programs accepted by the compiler.

For details, see the extensive commentary in Laziness.hs.

This is unlikely to break a working program, but the upshot for users is
that it's now possible to get a run-time error when dereferencing an
identifier in a recursive binding group before it has been initialized,
instead of silently getting an `undefined` value and having that maybe
or maybe not lead to an error somewhere else.