Add Prim.Int.ToString compiler-solved class (#4267)
* Add Prim.Int.ToString compiler-solved class

This class is being added primarily because large integers
outside the range of JavaScript's Int bounds can still be
reflected to the value level if the type used is
a BigInt rather than an Int. This is done by
converting a type-level Int to a type-level
String and then reflecting that String into
a value-level one that is used in BigInt's constructor.