ci/build-package-set.sh: exclude metadata package
`metadata` is a dummy package, not an actual part of the package set.
However, if asked, spago will try to install it anyway, and complain if
a matching version doesn't exist. When developing a new version of
PureScript, there might be a delay before the package set maintainers
get around to ensuring such a version exists. But there's no need to
wait for that if we simply don't install that package.