Fix bad interaction between module renaming and inliner
This bug was triggered when modules that the compiler handles specially
are shadowed by local constructors. For example, a constructor named
`Prim` could have caused references to `Prim_1["undefined"]` to be
produced in the compiled code, leading to a reference error at run time.
Less severely, a constructor named `Control_Bind` would have caused the
compiler not to inline known monadic functions, leading to slower and
less readable compiled code.