Float compiler-synthesized function applications (#3915)
This commit implements a common subexpression elimination pass during
CoreFn optimization. The optimizer only targets `App`s that the compiler
has created, meaning: applying functions to typeclass dictionaries,
accessing superclasses from a typeclass dictionary, or creating new
instances of `IsSymbol`. Furthermore, the pass only moves subexpressions
if they can be moved to an outer function or if the subexpression
appears more than once inside the scope to which it would be moved.