Enable more warnings in purescript.cabal
This enables -Wincomplete-uni-patterns and -Wincomplete-record-updates
by default. They are both good warnings to have on, as they help
identify situations where code is likely to result in an error at
runtime.

-Wincomplete-uni-patterns causes a warning to be emitted for things like
this:

    head = \(x:_) -> x

which would blow up at runtime if given an empty list.

-Wincomplete-record-updates causes a warning to be emitted for code like
this:

    data Foo = Foo { x :: Int }
             | Bar

    f :: Foo -> Foo
    f foo = foo { x = 6 }

which will blow up if you try to call `f Bar`, since only the `Foo`
constructor has the `x` field. (This one isn't triggered by any of our
code since we generally avoid incomplete record updates anyway).

The reason I'm sending this now is that as of version 9.2.1, GHC
includes these two warnings as part of -Wall by default, which means
that in order to upgrade, we'd need to either rework our code to stop
emitting them, or explicitly turn them off. The former seems more
appealing to me.

Another nice consequence of this is that some functions become more
efficient, as we remove unnecessary extra traversals of lists in a
couple of cases.