#! must be first chars in shebang
Previous implementation reused
much of the `breakComments` implementation.
However, that implementation allows the file to start
with spaces and newlines before
any comments are actually lexed.
Shebangs cannot have any spaces
or newlines preceding it. `#!` must be
the first chars in the file.

So, the implementation has changed to
reflect that. The grammar is now:

shebang :=
  | '#!' <text> <lineFeed> <shebang>
  | '#!' <text>