Drop UnusableDeclaration; update NoInstanceFound error (#4513)
With the advent of visible type applications (VTAs), all type class instances 
can be determined via VTAs. This PR removes the `UnusableDeclaration` error. 
Whereas before, the code would throw the above error when type variables 
in the type class head are not mentioned in the type class member's type 
signature, now we simply track these "VTA-required" args. 
These args represent type variables that must be disambiguated using VTAs.
If VTAs are used, the instance should be found. If they are not used,
an updated `NoInstanceFound` error is thrown that notifies the user of these
"VTA-required" args if possible.

When a `NoInstanceFound` error is thrown, we do not attempt to do anything
other than notify the user of the "VTA-required" args. In a previous and 
problematic approach, we did try to do more but found a number of difficult
problems with this approach:
- Suggesting a correct usage of `coerce` with VTAs in all of its complex cases
- When a user does supply a VTA arg, assuming that that arg refers to a corresponding 
    type in a valid type class instance: `foo @Int` when there is no `Foo Int` instance
- User confusion due to a difference in order of VTA args with their usage within 
    the type class member's type signature: `foo @1 @2` when `foo :: forall one two. two -> one`

A side effect of dropping the `UnusableDeclaration` error is a scoping bug 
that was revealed in `moveQuantifiersToFront`. This bug has also been fixed.

Other notes:
- `replaceAllTypeSynonymsM` is removed as it's no longer used by anything.
- The Externs files do not store the "VTA-required" args and instead recompute them after being loaded.
- The "VTA-required" args are tracked via their indices so as to print them in their corresponding order later.