Fix space leak between modules during compilation
For builds with a lot of warnings, memory usage grows drastically
since it appears that the thunks for the warnings hang onto a lot
of memory from compiling the module itself.

The goal of this change is to get memory usage for full builds
back in line with partial builds.